LayoutTests/fast/forms/validation-message-detached-iframe-expected.txt

Tests that the form validation message goes away when its element's iframe gets detached.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS message is ""
PASS successfullyParsed is true

TEST COMPLETE
